2st update

+++ APIs/Controllers/ServicesController.cs
@@ -6,6 +6,8 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using NuGet.Protocol.Plugins;
+using PetHealthcareSystem.APIs.DTOS;
 using PetHealthcareSystem.Models;
 using PetHealthcareSystem.Services;
 
@@ -46,16 +48,27 @@ namespace PetHealthcareSystem.Controllers
         // PUT: api/Services/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        public IActionResult UpdateService(Service toUpdateService)
+        public IActionResult UpdateService([FromRoute] int id, [FromBody] HealthServiceDTO toUpdateService)
         {
-            _healthService.UpdateHealthService(toUpdateService);
+            var service = _healthService.GetHealthServiceByCondition(s => s.ServiceId == id);
+            if (service == null)
+            {
+                return BadRequest();
+            }
+            Service updateService = new Service
+            {
+                ServicePrice = toUpdateService.ServicePrice,
+                ServiceId = id,
+                ServiceName = toUpdateService.ServiceName,
+            };
+            _healthService.UpdateHealthService(id, toUpdateService);
             return Ok(toUpdateService);
         }
 
         // POST: api/Services
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
-        public ActionResult<Service> CreateService(Service toUpdateService)
+        public ActionResult<Service> CreateService([FromBody] HealthServiceDTO toUpdateService)
         {
             _healthService.CreateHealthService(toUpdateService);
 
@@ -64,11 +77,15 @@ namespace PetHealthcareSystem.Controllers
 
         // DELETE: api/Services/5
         [HttpDelete("{id}")]
-        public IActionResult DeleteService(Service toDeleteService)
+        public IActionResult DeleteService([FromRoute] int id)
         {
+            var toDeleteService = _healthService.GetHealthServiceByCondition(s => s.ServiceId == id);
+            if(toDeleteService == null)
+            {
+                return NotFound(new {message ="Service not found"});
+            }
             _healthService.DeleteHealthService(toDeleteService);
-
-            return NoContent();
+            return Ok(toDeleteService);
         }
 
         private bool ServiceExists(int id)
